# docs/api_spec.yaml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Nordend"
  contact:
    name: Amagi Nordend Team
    email: "stv-dev@amagi.com"
  description: |-
    <br>
    Nordend - A service which holds the record of the master metadata belonging to an Account, Channel, Delivery and Platform that Amagi offers.

    <br/>
    <hr/>
    <br/>

    ### Entities

    - **Account** : Account is a resource that is uniquely identifiable with AMGID and represents one billable customer.
    - **Channel** : Channel is a resource that represents any linear delivery specific workflow that is running within amagi ecosystem.
    - **Platform** : Platform is a resource that uniquely refers to any consumers of the deliverables that Amagi has to offer via linear or on-demand workflows.
    - **Delivery** : Delivery is an association entity which connects a account or a channel to one platform.

    <br/>
    <hr/>
    <br/>
    
    ### Entity Relationship

    | Entity | Association | Entity |
    |-|-|-|
    | Account | One to Many | Channel |
    | Channel | One to Many | Delivery |
    | Delivery | One to One | Platform |

    <br/>
    <hr/>
    <br/>
    
    ### API Usage

    #### Authentication

    - API Key Authentitcation is used to make any API calls.
    - API Key can be obtained by contacting the developer team .
    - API Key needs to be passed alongside the API authorization header for every API request.

    #### Response Codes

    | HTTP Verb | Allowed Responses |
    | - | - |
...

...
servers:
...
paths:
...
  /channels/{channelid}:
    get:
      tags:
        - "Channel Management"
      summary: "Get details of a specific channel"
      description: |-
        This API allows the caller to get a specific deatils of a channel resource.
      operationId: "getSpecificChannel"
      parameters:
        - in: "path"
          name: "channelid"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        403:
          description: "Unauthorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unauthorized"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"

    patch:
      tags:
        - "Channel Management"
      summary: "Update speific details of a Channel"
      description: |-
        This API allows the caller to change any specific values of the channel resource except audios, captions and artworks.

        This patch follows the implementation as specified in [RFC 6902](https://www.rfc-editor.org/rfc/rfc6902)

        <hr/>
        
        `add`

        The "add" operation performs one of the following functions, depending upon what the target location references
        - If the target location specifies an array index, a new value is inserted into the array at the specified index.
        - If the target location specifies an object member that does not already exist, a new member is added to the object.
        - If the target location specifies an object member that does exist, that member's value is replaced.

        Example: 
        ```
        { "op": "add", "path": "/name", "value": "mars" }
        ```
        <hr/>
        
        `remove`

        - The "remove" operation removes the value at the target location.
        - The target location MUST exist for the operation to be successful.

        Example:
        ```
        { "op": "remove", "path": "/name" }
        ```
        <hr/>
        
        `replace`

        - The "replace" operation replaces the value at the target location with a new value. The operation object MUST contain a "value" member whose content specifies the replacement value.
        - The target location MUST exist for the operation to be successful.

        Example: 
        ```
        { "op": "replace", "path": "/short_description", "value": "This is an example" }
        ```
        <hr/>
        
        `move`

        - The "move" operation removes the value at a specified location and adds it to the target location.
        - The operation object MUST contain a "from" member, which is a string containing a JSON Pointer value that references the location in the target document to move the value from.
        - The "from" location MUST exist for the operation to be successful.

        Example: 
        ```
        { "op": "move", "from": "/short_description", "path": "/long_description" }
        ```
        <hr/>

        `copy`

        - The "copy" operation copies the value at a specified location to the target location.
        - The operation object MUST contain a "from" member, which is a string containing a JSON Pointer value that references the location in the target document to copy the value from.
        - The "from" location MUST exist for the operation to be successful.

        Example: 
        ```
        { "op": "copy", "from": "/short_description", "path": "/long_description" }
        ```
      parameters:
        - in: "path"
          name: "channelid"
          required: true
          schema:
            type: "string"
      operationId: "updateChannel"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelPatch"
      responses:
        200:
          description: "successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelResponse"
        400:
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequest"
        403:
          description: "User forbidden from making the request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forbidden"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"

    delete:
      tags:
        - "Channel Management"
      summary: "Delete a Channel"
      description: |-
        This API allows the caller to delete specific channel resource.
        
        Note: 
          - This is an irrevocable action. Ensure you are absolutely sure before this is called.
          - This can be called, iff all the associated delivery resources are deleted.

      parameters:
        - in: "path"
          name: "channelid"
          required: true
          schema:
            type: "string"
      operationId: "deleteChannel"
      responses:
        "200":
          description: "successful operation"
        403:
          description: "User forbidden from making the request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Forbidden"
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
  schemas:
    BadRequest:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Invalid Request"
    Unauthorized:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Unauthorized"
    Forbidden:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Forbidden"
...
    InternalServerError:
      type: "object"
      properties:
        errors:
          type: "object"
        message:
          type: "string"
          example: "Internal Server Error"
...
    AudioLanguages:
      type: "string"
      enum:
            [
              "ABK",
              "AAR",
              "AFR",
              "AKA",
              "SQI",
              "AMH",
              "ARA9",
              "ARG",
              "HYE",
              "ASM",
              "AVA",
              "AVE",
              "AYM",
              "AZE",
              "BAM",
              "BAK",
              "EUS",
              "BEL",
              "BEN",
              "BIS",
              "BOS",
              "BRE",
              "BUL",
              "MYA",
              "CAT",
              "CHA",
              "CHE",
              "NYA",
              "ZHO6",
              "CHV",
              "COR",
              "COS",
              "CRE",
              "HRV",
              "CES",
              "DAN",
              "DIV",
              "NLD",
              "DZO",
              "ENG",
              "EPO",
              "EST",
              "EWE",
              "FAO",
              "FIJ",
              "FIN",
              "FRA",
              "FUL",
              "GLG",
              "KAT",
              "DEU",
              "ELL",
              "GRN",
              "GUJ",
              "HAT",
              "HAU",
              "HEB",
              "HER",
              "HIN",
              "HMO",
              "HUN",
              "INA",
              "IND",
              "ILE",
              "GLE",
              "IBO",
              "IPK",
              "IDO",
              "ISL",
              "ITA",
              "IKU",
              "JPN",
              "JAV",
              "KAL",
              "KAN",
              "KAU",
              "KAS",
              "KAZ",
              "KHM",
              "KIK",
              "KIN",
              "KIR",
              "KOM",
              "KON",
              "KOR",
              "KUR",
              "KUA",
              "LAT",
              "LTZ",
              "LUG",
              "LIM",
              "LIN",
              "LAO",
              "LIT",
              "LUB",
              "LAV",
              "GLV",
              "MKD",
              "MLG1",
              "MSA6",
              "MAL",
              "MLT",
              "MRI",
              "MAR",
              "MAH",
              "MON",
              "NAU",
              "NAV",
              "NDE",
              "NEP",
              "NDO",
              "NOB",
              "NNO",
              "NOR",
              "III",
              "NBL",
              "OCI",
              "OJI",
              "CHU",
              "ORM",
              "ORI",
              "OSS",
              "PAN",
              "PLI",
              "FAS",
              "POL",
              "PUS",
              "POR",
              "QUE3",
              "ROH",
              "RUN",
              "RON",
              "RUS",
              "SAN",
              "SRD",
              "SND",
              "SME",
              "SMO",
              "SAG",
              "SRP",
              "GLA",
              "SNA",
              "SIN",
              "SLK",
              "SLV",
              "SOM",
              "SOT",
              "SPA",
              "SUN",
              "SWA",
              "SSW",
              "SWE",
              "TAM",
              "TEL",
              "TGK",
              "THA",
              "TIR",
              "BOD",
              "TUK",
              "TGL",
              "TSN",
              "TON",
              "TUR",
              "TSO",
              "TAT",
              "TWI",
              "TAH",
              "UIG",
              "UKR",
              "URD",
              "UZB",
              "VEN",
              "VIE",
              "VOL",
              "WLN",
              "CYM",
              "WOL",
              "FRY",
              "XHO",
              "YID",
              "YOR",
              "ZHA6",
              "ZUL",
            ]
    Audios:
      type: "array"
      items:
        $ref: "#/components/schemas/AudioLanguages"
...
    Artwork:
      type: "object"
      properties:
        width:
          type: "number"
          example: "1920"
        height:
          type: "number"
          example: "1080"
        file:
          type: "string"
        labels:
          type: "object"
    Caption:
      type: "array"
      items:
        $ref: "#/components/schemas/CaptionObject"
    CaptionObject:
      type: "object"
      properties:
        format:
          type: "string"
          enum: ["webVTT", "srt", "scc"]
          example: "srt"
        language:
          $ref: "#/components/schemas/Audios"
    Audio:
      type: "object"
      properties:
        track:
          type: "number"
          example: 1
          enum: [1, 2, 3, 4, 5, 6, 7, 8]
        language:
          $ref: "#/components/schemas/AudioLanguages"
        is_primary:
          type: "boolean"
          example: true
    Timezone:
      type: "string"
      enum:
        [
          "International Date Line West",
          "American Samoa",
          "Midway Island",
          "Hawaii",
          "Alaska",
          "Pacific Time (US &amp; Canada)",
          "Arizona",
          "Tijuana",
          "Chihuahua",
          "Mazatlan",
          "Mountain Time (US &amp; Canada)",
          "Central America",
          "Saskatchewan",
          "Guadalajara",
          "Mexico City",
          "Monterrey",
          "Central Time (US &amp; Canada)",
          "Bogota",
          "Quito",
          "Lima",
          "Indiana (East)",
          "Eastern Time (US &amp; Canada)",
          "Puerto Rico",
          "La Paz",
          "Georgetown",
          "Caracas",
          "Atlantic Time (Canada)",
          "Santiago",
          "Montevideo",
          "Buenos Aires",
          "Brasilia",
          "Newfoundland",
          "Greenland",
          "Mid-Atlantic",
          "Cape Verde Is.",
          "UTC",
          "Azores",
          "Monrovia",
          "London",
          "Lisbon",
          "Edinburgh",
          "West Central Africa",
          "Casablanca",
          "Dublin",
          "Bern",
          "Bratislava",
          "Brussels",
          "Budapest",
          "Berlin",
          "Copenhagen",
          "Belgrade",
          "Ljubljana",
          "Madrid",
          "Paris",
          "Prague",
          "Rome",
          "Sarajevo",
          "Skopje",
          "Stockholm",
          "Vienna",
          "Warsaw",
          "Amsterdam",
          "Zagreb",
          "Harare",
          "Pretoria",
          "Kaliningrad",
          "Cairo",
          "Zurich",
          "Helsinki",
          "Jerusalem",
          "Bucharest",
          "Kyiv",
          "Athens",
          "Riga",
          "Sofia",
          "Volgograd",
          "Vilnius",
          "Baghdad",
          "Istanbul",
          "Kuwait",
          "Minsk",
          "Moscow",
          "Nairobi",
          "Riyadh",
          "St. Petersburg",
          "Tallinn",
          "Tehran",
          "Abu Dhabi",
          "Baku",
          "Muscat",
          "Samara",
          "Yerevan",
          "Tbilisi",
          "Kabul",
          "Islamabad",
          "Ekaterinburg",
          "Tashkent",
          "Karachi",
          "Chennai",
          "Kolkata",
          "Mumbai",
          "New Delhi",
          "Sri Jayawardenepura",
          "Kathmandu",
          "Astana",
          "Urumqi",
          "Dhaka",
          "Almaty",
          "Rangoon",
          "Jakarta",
          "Hanoi",
          "Bangkok",
          "Krasnoyarsk",
          "Novosibirsk",
          "Chongqing",
          "Beijing",
          "Hong Kong",
          "Irkutsk",
          "Ulaanbaatar",
          "Perth",
          "Singapore",
          "Taipei",
          "Kuala Lumpur",
          "Osaka",
          "Sapporo",
          "Seoul",
          "Tokyo",
          "Yakutsk",
          "Darwin",
          "Brisbane",
          "Guam",
          "Port Moresby",
          "Vladivostok",
          "Adelaide",
          "Hobart",
          "Canberra",
          "Sydney",
          "Melbourne",
          "Magadan",
          "Srednekolymsk",
          "Solomon Is.",
          "New Caledonia",
          "Marshall Is.",
          "Fiji",
          "Kamchatka",
          "Auckland",
          "Wellington",
          "Tokelau Is.",
          "Samoa",
          "Nuku&#39;alofa",
          "Chatham Is",
        ]
      example: "International Date Line West"
    ChannelPatch:
      type: "object"
      properties:
        op:
          type: "string"
          enum: ["add", "remove", "replace", "move", "copy"]
        path:
          type: "string"
          enum: ["/name", "/short_description", "/long_description", "/start_time", "/timezone", "/genre", "/tags"]
        value:
          type: "string"
          example: "Mars"
...
    Rating:
      type: "object"
      properties:
        body:
          type: "string"
        value:
          type: "string"
          example: "PG 16"
    ChannelResponse:
      type: "object"
      properties:
        channelid:
          type: "string"
          example: AMG00001C1
        amgid:
          type: "string"
          example: AMG00001
        name:
          type: "string"
          example: "Nebula"
        short_description:
          type: "string"
          example: "The wonders of galaxy."
        long_description:
          type: "string"
          example: "Showcasing the wonderful galaxy through the eyes of the human."
        start_time:
          type: "string"
          example: "18:00"
        timezone:
          $ref: "#/components/schemas/Timezone"
        rating:
          type: "array"
          items:
            $ref: "#/components/schemas/Rating"
        genre:
          type: "string"
          example: "action,comedy"
          enum: ["action", "adventure", "comedy", "drama", "fantasy", "horror", "musicals", "mystery", "romance", "science fiction", "sports", "thriller", "western"]
        audios:
          type: "array"
          items:
            $ref: "#/components/schemas/Audio"
        caption:
          $ref: "#/components/schemas/Caption"
        artworks:
          type: "array"
          items:
            $ref: "#/components/schemas/Artwork"
        status:
          type: "string"
          example: "success"
          enum: ["success", "failed", "starting", "provisioning"]
        tags:
          type: "object"
...
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
security:
  - token: []